Handled bkcharts compatibility
